+++ src/Defender.Model/Defender.Model.Base.cs
@@ -6,7 +6,7 @@
     using System.Text;
     using System.Threading.Tasks;
 
-    public class ModelBase
+    public static class ModelBase
     {
     }
 }
